requirements.txt
PySide6
cffi
pillow